requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
montydb==2.5.3
Werkzeug==3.0.0
flask-pymongo==3.0.1